requirements.txt
dgsamples==1.1.2
Flask==1.1.2
GDAL==2.3.2
geoio==1.3.0
h5py==2.8.0
imutils==0.5.4
Keras==2.2.4
Keras_Applications==1.0.8
matplotlib==2.0.2
numpy==1.19.1
opencv_python==3.4.2.16
osgeo==0.0.1
pandas==1.1.1
Pillow==8.4.0
pyproj==1.9.5.1
pyshp==2.1.3
rasterio==1.2.0
requests==2.25.1
scikit_image==0.16.2
scikit_learn==1.0.2
scipy==1.5.2
segmentation_models==1.0.1
skimage==0.0
tensorflow==1.12.0
tifffile==2021.11.2
